+++ packages/app/src/utils/network.ts
@@ -5,5 +5,5 @@ export const correctChainIsLoaded = async (pepemon: any) => {
     }
 
     const { chainId } = await pepemon.provider.getNetwork();
-    return pepemon.chainId === parseInt(chainId);
+    return pepemon.chainId === parseFloat(chainId);
 }
